Splitting checks between general MIP CV and specific CORDEX-CMIP6 checks Adding download of miptables similar to euro-cordex/cmor-check
