+++ cc_plugin_cc6/tables.py
@@ -1,0 +1,11 @@
+def retrieve(url, fname, path):
+    import pooch
+
+    # URL to one of Pooch's test files
+    filename = pooch.retrieve(
+        url=url,
+        fname=fname,
+        known_hash=None,
+        path=path,
+    )
+    return filename
